Massively improved numerical stability and added adaptive supersampling
I identified the main issue with the numerical stability: the scale of the input
data essentially led to such bas floating point errors that it actually improved
the results for edge cases where a ray lies normal to the triangle it is
testing. In other cases, it made the results worse.
Evetually, the best solution will be to normalize the vertices to $[-1, 1]$.
Furthermore, regarding the sampling, I opted for an adaptive sampling solution,
where so long as the relative differences between subpixels are not small
enough, we keep subidivding, using a quadtree scheme.